requirements.txt
pydantic-ai>=0.0.30
langgraph>=0.3.1
langchain-community>=0.3.18
langchain>=0.3.19
langchain_google_genai>=2.0.10
logfire>=3.6.4
python-dotenv>=1.0.1
google-generativeai>=0.8.0 